requirements.txt
absl-py==0.13.0
astor==0.8.1
astunparse==1.6.3
audeer==1.16.0
audformat==0.11.3
audinterface==0.6.6
audiofile==0.4.2
audobject==0.4.13
audresample==0.1.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
importlib-metadata==4.6.0
iso-639==0.4.5
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opensmile==2.1.2
opt-einsum==3.3.0
oyaml==1.0
pandas==1.2.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SoundFile==0.10.3.post1
sox==1.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tqdm==4.61.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
